Update zshenv and zshrc

+++ zsh/zsh_install.sh
@@ -1,8 +1,17 @@
 #!/bin/sh
-sudo apt install zsh
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if ! type zsh &> /dev/null; then
+  sudo apt install zsh
+  sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+fi
 
 DOTFILES_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-ln -sf "$DOTFILES_DIR/zshrc" $HOME/.zshrc
-source $HOME/.zshrc
+if [ ! -e "$HOME/.zshrc" ] ; then
+  ln -sf "$DOTFILES_DIR/zshrc" $HOME/.zshrc
+  source $HOME/.zshrc
+fi
+
+if [ ! -e "$HOME/.zshenv" ] ; then
+  ln -sf "$DOTFILES_DIR/zshenv" $HOME/.zshenv
+  source $HOME/.zshenv
+fi
